ratsd-token-v2: initialize CBOR modes in init
